學員作品/信用卡消費樣態/flask_dash_web/requirements.txt
requests
pandas
numpy
matplotlib
seaborn
scikit-learn
flask
dash
gunicorn
dash-bootstrap-components
psycopg2-binary
nbformat